backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
python-multipart==0.0.9
python-docx==1.1.2
reportlab==4.2.2
openpyxl==3.1.5
pandas==2.2.2
pydantic==2.9.0
aiofiles==24.1.0
